pkg: reorganize helpers.
