+++ lib/device/helpers/common.js
@@ -7,8 +7,8 @@
 
 const assert = require('bsert');
 const {BufferMap} = require('buffer-map');
-const {Path} = require('../path');
-const {InputData} = require('../inputData');
+const {Path} = require('../../path');
+const {InputData} = require('../../inputData');
 
 const helpers = exports;
