+++ lib/device/helpers/ledger.js
