+++ lib/device/helpers/trezor.js
