+++ lib/device/ledger.js
@@ -11,8 +11,8 @@ const bledger = require('bledger');
 const AbstractDeviceWrapper = require('./abstract');
 const {vendors} = require('../common');
 const MTX = require('bcoin/lib/primitives/mtx');
-const common = require('./helpers');
-const helpers = require('./ledgerhelpers');
+const common = require('./helpers/common');
+const helpers = require('./helpers/ledger');
 
 const {
   ManagedLedgerBcoin,
