+++ lib/device/trezor.js
@@ -14,8 +14,8 @@ const HDPublicKey = require('bcoin/lib/hd/public');
 const MTX = require('bcoin/lib/primitives/mtx');
 const {vendors} = require('../common');
 const AbstractDeviceWrapper = require('./abstract');
-const common = require('./helpers');
-const helpers = require('./trezorhelpers');
+const common = require('./helpers/common');
+const helpers = require('./helpers/trezor');
 
 class TrezorDeviceWrapper extends AbstractDeviceWrapper {
   constructor(options) {
