+++ lib/internal/usb-browser.js
