+++ lib/internal/usb.js
