+++ lib/manager/ledger.js
@@ -8,7 +8,7 @@
 const assert = require('bsert');
 const Network = require('bcoin/lib/protocol/network');
 const busb = require('busb');
-const {getUSB} = require('./usb');
+const {getUSB} = require('../internal/usb');
 const LedgerUSB = require('bledger/lib/device/usb').Device;
 const LedgerDeviceWrapper = require('../device/ledger');
 const AbstractDeviceManager = require('./abstract');
