+++ lib/manager/manager.js
@@ -22,7 +22,7 @@ const VENDOR_MANAGERS = {
 /**
  * Device Manager
  *
- * TODO: Add getDevices and deviceSelector callback.
+ * TODO: Add getDevices.
  */
 class DeviceManager extends EventEmitter {
   constructor(options) {
@@ -183,6 +183,17 @@ class DeviceManager extends EventEmitter {
   getPublicKey(path, getParentFingerPrint = true) {
     assert(this.selected, 'Device was not selected.');
     return this.selected.getPublicKey(path, getParentFingerPrint);
+  }
+
+  /**
+   * Get public key in xpub string format.
+   * @param {Path} path
+   * @returns {Promise<string>} - xpub
+   */
+
+  async getXPUB(path) {
+    assert(this.selected, 'Device was not selected.');
+    return this.selected.getXPUB(path);
   }
 
   /**
